+++ src/Vsix/Classifier/EditorConfigClassifier.cs
@@ -26,7 +26,7 @@ namespace EditorConfig
                     {Tuple.Create(_rxHeader, registry.GetClassificationType(PredefinedClassificationTypeNames.String))},
                     {Tuple.Create(_rxKeyword, registry.GetClassificationType(PredefinedClassificationTypeNames.Keyword))},
                     {Tuple.Create(_rxValue, registry.GetClassificationType(PredefinedClassificationTypeNames.SymbolDefinition))},
-                    {Tuple.Create(_rxSeverity, registry.GetClassificationType(PredefinedClassificationTypeNames.SymbolReference))},
+                    {Tuple.Create(_rxSeverity, registry.GetClassificationType(PredefinedClassificationTypeNames.Identifier))},
                 };
         }