+++ src/Vsix/Constants.cs
@@ -5,5 +5,7 @@
         public const string LanguageName = "EditorConfig";
         public const string FileName = ".editorconfig";
         public const string Homepage = "http://editorconfig.org";
+
+        public static string[] Severity { get; } = { "none", "suggestion", "warning", "error" };
     }
 }